# backend/docs/swagger.yaml
basePath: /
definitions:
  domain.Role:
    enum:
    - CANDIDATE
    - RECRUITER
    type: string
    x-enum-varnames:
    - RoleCandidate
    - RoleRecruiter
  dto.ApplyJobOutputDTO:
    properties:
      applied_at:
        type: string
      candidate_id:
        type: integer
      candidate_name:
        type: string
      company:
        type: string
      id:
        type: integer
      job_id:
        type: integer
      job_title:
        type: string
      location:
        type: string
      status:
        type: string
    type: object
  dto.CreateJobOutputDTO:
    properties:
      anonymous:
        type: boolean
      company:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      recruiter_email:
        type: string
      recruiter_id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  dto.DashboardStatsDTO:
    properties:
      applied:
        type: integer
      pending:
        type: integer
    type: object
  dto.GetJobOutputDTO:
    properties:
      anonymous:
        type: boolean
      company:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      recruiter_email:
        type: string
      recruiter_id:
        type: integer
      requirements:
        type: string
      salary:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  dto.MetaDTO:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.PaginatedApplicationsOutputDTO:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ApplyJobOutputDTO'
        type: array
      meta:
        $ref: '#/definitions/dto.MetaDTO'
    type: object
  dto.PaginatedJobsOutputDTO:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.GetJobOutputDTO'
        type: array
      meta:
        $ref: '#/definitions/dto.MetaDTO'
    type: object
  dto.RegisterOutputDTO:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
    type: object
  web.CreateJobRequest:
    properties:
      anonymous:
        type: boolean
      company:
        type: string
      description:
        type: string
      location:
        type: string
      requirements:
        type: string
      salary:
        type: string
      title:
        type: string
    required:
    - company
    - description
    - location
    - title
    type: object
  web.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  web.FinalizeJobRequest:
    properties:
      candidate_id:
        type: integer
    required:
    - candidate_id
    type: object
  web.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  web.LoginResponse:
    properties:
      token:
        type: string
    type: object
  web.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        allOf:
        - $ref: '#/definitions/domain.Role'
        enum:
        - CANDIDATE
        - RECRUITER
    required:
    - email
    - name
    - password
    - role
    type: object
  web.UpdateJobRequest:
    properties:
      company:
        type: string
      description:
        type: string
      location:
        type: string
      requirements:
        type: string
      salary:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend for Recruitment and Selection System
  title: Recruitment System API
  version: "1.0"
paths:
  /applications:
    get:
      consumes:
      - application/json
      description: List all applications for the logged in candidate
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedApplicationsOutputDTO'
      security:
      - BearerAuth: []
      summary: Get my applications
      tags:
      - applications
  /applications/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel a pending application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel an application
      tags:
      - applications
  /dashboard/summary:
    get:
      consumes:
      - application/json
      description: Get statistics for the dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardStatsDTO'
      security:
      - BearerAuth: []
      summary: Get dashboard summary
      tags:
      - dashboard
  /jobs:
    get:
      consumes:
      - application/json
      description: Get all jobs with optional search query and pagination
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Status filter (OPEN|PAUSED|CLOSED)
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedJobsOutputDTO'
      summary: List all jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Create a job posting (Recruiter only)
      parameters:
      - description: Create Job Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateJobOutputDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new job
      tags:
      - jobs
  /jobs/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetJobOutputDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Get job by ID
      tags:
      - jobs
    patch:
      consumes:
      - application/json
      description: Update job fields; only OPEN or PAUSED jobs can be updated. Status
        can be OPEN or PAUSED.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Job Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.UpdateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetJobOutputDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a job (Recruiter only)
      tags:
      - jobs
  /jobs/{id}/applications:
    get:
      consumes:
      - application/json
      description: List all applications for a specific job (Recruiter only)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status (e.g., PENDING)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedApplicationsOutputDTO'
      security:
      - BearerAuth: []
      summary: Get job applications
      tags:
      - applications
  /jobs/{id}/apply:
    post:
      consumes:
      - application/json
      description: Apply for a job as a candidate
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ApplyJobOutputDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Apply for a job
      tags:
      - applications
  /jobs/{id}/finalize:
    post:
      consumes:
      - application/json
      description: Close the job and mark the specified candidate as hired
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Finalize Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.FinalizeJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Finalize a job and hire a candidate
      tags:
      - jobs
  /jobs/mine:
    get:
      consumes:
      - application/json
      description: Get jobs created by the logged-in recruiter with optional search
        and pagination
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Status filter (OPEN|PAUSED|CLOSED)
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedJobsOutputDTO'
      security:
      - BearerAuth: []
      summary: List recruiter-owned jobs
      tags:
      - jobs
  /login:
    post:
      consumes:
      - application/json
      description: Login and get JWT token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterOutputDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Register a new user (Candidate or Recruiter)
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"